wallet_server/requirements.txt
Django==1.10.8
djangorestframework==3.6.4
MySQL-python==1.2.5
